% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{visu_accumulate}
\alias{visu_accumulate}
\title{Kerää lyhyen ikkunan rajapinnasta pitkä sarja}
\usage{
visu_accumulate(data, path)
}
\arguments{
\item{data}{Uudet havainnot. Sarake \code{values} on arvo, muut sarakkeet
yhdessä yksilöivät havainnon.}

\item{path}{Csv-tiedoston polku. Luodaan jos sitä ei vielä ole.}
}
\value{
Koko kertynyt sarja data framena.
}
\description{
Yhdistää uudet havainnot aiemmin tallennettuun csv-tiedostoon ja palauttaa
koko kertyneen sarjan. Tarkoitettu lähteille, jotka näyttävät vain
viimeisimmät viikot: kun sivusto ajetaan säännöllisesti, historia karttuu
tiedostoon eikä katkea.
}
\details{
Päällekkäisissä havainnoissa uusi arvo voittaa, jotta lähteen korjaukset
menevät läpi. Tiedosto kirjoitetaan aina järjestyksessä, jotta git-diff
näyttää vain uudet rivit.

Jos haku epäonnistuu, funktio palauttaa aiemmin kertyneen sarjan ja
varoittaa. Näin yhden rajapinnan hetkellinen katko jättää kuvion
edellisiin havaintoihin sen sijaan että se kaataisi koko renderöinnin.
Tiedosto kirjoitetaan vain kun uusia havaintoja saatiin.
}
\examples{
\dontrun{
visu_get_bof("https://www.suomenpankki.fi/api/interestrates/euribor") |>
  visu_accumulate("../data/euribor.csv")
}
}
