% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{visu_catalog_path}
\alias{visu_catalog_path}
\title{Kuvioluettelon tiedosto}
\usage{
visu_catalog_path(site_dir = NULL)
}
\arguments{
\item{site_dir}{Sivuston hakemisto, ks. \code{\link[=visu_site_dir]{visu_site_dir()}}.}
}
\value{
Tiedostopolku merkkijonona.
}
\description{
Kuvioluettelon tiedosto
}
