% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshness.R
\name{visu_clear_cache}
\alias{visu_clear_cache}
\title{Tyhjennä kansiolistausten välimuisti}
\usage{
visu_clear_cache()
}
\value{
\code{NULL}, kutsutaan sivuvaikutuksen vuoksi.
}
\description{
Tyhjennä kansiolistausten välimuisti
}
