% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{visu_site_dir}
\alias{visu_site_dir}
\title{Sivuston lähdehakemisto}
\usage{
visu_site_dir(site_dir = NULL)
}
\arguments{
\item{site_dir}{Hakemisto, tai \code{NULL} jolloin käytetään asetusta
\code{visu.site_dir} ja sen puuttuessa työhakemiston alla olevaa \code{site}-kansiota.}
}
\value{
Hakemistopolku merkkijonona.
}
\description{
Sivuston lähdehakemisto
}
