% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{visu_state_read}
\alias{visu_state_read}
\title{Lue kuvioiden tilatiedosto}
\usage{
visu_state_read(site_dir = NULL)
}
\arguments{
\item{site_dir}{Sivuston hakemisto, ks. \code{\link[=visu_site_dir]{visu_site_dir()}}.}
}
\value{
Nimetty lista, jonka avaimina ovat kuvioiden tunnisteet. Tyhjä lista
jos tiedostoa ei vielä ole.
}
\description{
Lue kuvioiden tilatiedosto
}
