% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{visu_state_write}
\alias{visu_state_write}
\title{Kirjoita kuvioiden tilatiedosto}
\usage{
visu_state_write(state, site_dir = NULL)
}
\arguments{
\item{state}{Nimetty lista, ks. \code{\link[=visu_state_read]{visu_state_read()}}.}

\item{site_dir}{Sivuston hakemisto.}
}
\value{
Tiedostopolku näkymättömänä.
}
\description{
Avaimet järjestetään aakkosjärjestykseen, jotta git-diff näyttää vain
oikeasti muuttuneet kuviot.
}
