% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu-package.R
\docType{package}
\name{visu-package}
\alias{visu}
\alias{visu-package}
\title{visu: StatFin Datan Haku ja Visualisointi}
\description{
Rakentaa StatFin-datasta interaktiivisia kuvioita ggplot2:lla ggcustom-teemalla ja plotlylla, ja julkaisee ne Quarto-sivustona. Sivusto paivittyy inkrementaalisesti: paketti tarkistaa kunkin kuvion lahdetaulun paivitysajan PxWeb-rajapinnasta ja renderoi uudelleen vain ne kuviot, joiden data on muuttunut.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jhuovari.github.io/visu/}
  \item \url{https://github.com/jhuovari/visu}
  \item Report bugs at \url{https://github.com/jhuovari/visu/issues}
}

}
\author{
\strong{Maintainer}: Janne Huovari \email{janne.huovari@gov.fi}

}
\keyword{internal}
